requirements.txt
# Core Framework
streamlit==1.23.1

# AI/ML APIs
openai>=1.0.0
anthropic>=0.18.0

# Audio Processing
pydub==0.25.1

# Environment & Configuration
python-dotenv==1.0.0

# Natural Language Processing
spacy==3.5.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.0

# Language Detection & Translation
langdetect==1.0.9

# Text Processing & Summarization
bert-extractive-summarizer==0.10.1

# Database
psycopg2-binary==2.9.6
alembic==1.11.1

# WordPress Integration
python-wordpress-xmlrpc==2.3

# Analytics
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0

# Markdown Processing
markdown2==2.4.8

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Code Quality
black==23.3.0
flake8==6.0.0
pre-commit==3.3.2
